Move giving season constants into lib
