+++ packages/lesswrong/components/ea-forum/giving-portal/EAGivingPortalPage.tsx
@@ -4,8 +4,18 @@ import { AnalyticsContext } from "../../../lib/analyticsEvents";
 import { Link } from "../../../lib/reactRouterWrapper";
 import { SECTION_WIDTH } from "../../common/SingleColumnSection";
 import { formatStat } from "../../users/EAUserTooltipContent";
-import { useDonationOpportunities, useElectionCandidates } from "./hooks";
-import type { TimelineSpec } from "./Timeline";
+import {
+  useAmountRaised,
+  useDonationOpportunities,
+  useElectionCandidates,
+} from "./hooks";
+import {
+  donationElectionLink,
+  donationElectionTagId,
+  effectiveGivingTagId,
+  timelineSpec,
+  votingOpensDate,
+} from "../../../lib/eaGivingSeason";
 import classNames from "classnames";
 
 const styles = (theme: ThemeType) => ({
@@ -168,50 +178,6 @@ const styles = (theme: ThemeType) => ({
   mb80: { marginBottom: 80 },
   mb100: { marginBottom: 100 },
 });
-
-const useAmountRaised = () => {
-  // TODO: Query for the actual amount
-  return {
-    raisedForElectionFund: 3720,
-    donationTarget: 15000,
-    totalRaised: 10250,
-  };
-}
-
-const donationElectionLink = "#"; // TODO
-
-const votingOpensDate = new Date("2023-12-01");
-
-const donationElectionTagId = "L6NqHZkLc4xZ7YtDr"; // TODO: This tag doesn't exist yet
-const effectiveGivingTagId = "L6NqHZkLc4xZ7YtDr";
-
-const timelineSpec: TimelineSpec = {
-  start: new Date("2023-11-15"),
-  end: new Date("2023-12-31"),
-  points: [
-    {date: new Date("2023-11-28"), description: "Giving Tuesday"},
-    {date: votingOpensDate, description: "Voting starts"},
-    {date: new Date("2023-12-15"), description: "Voting ends"},
-    {date: new Date("2023-12-20"), description: "Election winner announced"},
-  ],
-  spans: [
-    {
-      start: new Date("2023-11-21"),
-      end: new Date("2023-11-28"),
-      description: "Effective giving spotlight Week",
-    },
-    {
-      start: new Date("2023-11-30"),
-      end: new Date("2023-12-07"),
-      description: "Marginal Funding Week",
-    },
-    {
-      start: new Date("2023-12-08"),
-      end: new Date("2023-12-16"),
-      description: "Forum BOTEC-a-thon Week",
-    },
-  ],
-};
 
 const getListTerms = (
   tagId: string,
