+++ packages/lesswrong/components/ea-forum/giving-portal/hooks.ts
@@ -1,16 +1,24 @@
 import { useMulti } from "../../../lib/crud/withMulti";
-
-const givingSeason23ElectionName = "givingSeason23";
+import { eaGivingSeason23ElectionName } from "../../../lib/eaGivingSeason";
 
 export const useElectionCandidates = () => {
   return useMulti({
     collectionName: "ElectionCandidates",
     fragmentName: "ElectionCandidateBasicInfo",
     terms: {
-      electionName: givingSeason23ElectionName,
+      electionName: eaGivingSeason23ElectionName,
     },
   });
 }
 
 // TODO: Should this have separate logic to useElectionCandidates?
 export const useDonationOpportunities = useElectionCandidates;
+
+export const useAmountRaised = () => {
+  // TODO: Query for the actual amount
+  return {
+    raisedForElectionFund: 3720,
+    donationTarget: 15000,
+    totalRaised: 10250,
+  };
+}
